Make the Parallel Computing Toolbox optional for model computation
Currently, no parallel computing is used during model computation.
There are, however, structures in place for it to be used in the
future. Specifically, each level of the nested computation (validation,
testing, both folds and iterations) checks if a parfor loop is already
running. This check is necessary due to issues with nested parfor
loops. Currently, the result of the check will always be false because
no parfor loop is used. The check itself, however, involves a call to
getCurrentTask() which is only available through the Parallel Computing
Toolbox. Thus, no model can be computed without it.

This commit wraps each of these checks in a try/catch and returns false
if it fails. In this way, model computation is possible without PCT
installed.